feat: Update references to model parameters in preprocforVAE_T_Conv
This commit updates the references to model parameters in the `preprocforVAE_T_Conv.m` file. The `load` function now loads the `reference_frame` from the `Data/reference_frame.mat` file instead of the `Data/reference_frame_V2.mat` file. This change ensures that the correct reference frame is used for further processing.

It also updates the `getProcessedFile` function in the `Project.m` file to include a default value for the `processingType` parameter. This allows the function to be called without specifying the `processingType` argument.

Lastly, it updates the `reconstruct_sleap_preds` function in the `reconstruct_sleap_preds.m` file to use the `modelPath` parameter instead of the `path` parameter when getting predictions from the Python model. This change ensures that the correct model path is used for obtaining predictions.

These changes improve the accuracy and reliability of the pose reconstruction process.

And further changes showing how to use this in the howto.m