feat: Add functionality to return a list of processed files with default reconstruction method
This commit adds the `returnDefaultReconstructionFiles` function to the `Project.m` file. This function returns a cell array of all the files that have been processed with the default reconstruction method. If no reconstruction method is specified, it defaults to 'default_reconstruction'.

By providing this functionality, developers can easily retrieve a list of processed files for further analysis or processing. This improves the efficiency and flexibility of the codebase.